+++ protocols/horizon/main.go
@@ -432,26 +432,43 @@ type Transaction struct {
 		Precedes   hal.Link `json:"precedes"`
 		Succeeds   hal.Link `json:"succeeds"`
 	} `json:"_links"`
-	ID              string    `json:"id"`
-	PT              string    `json:"paging_token"`
-	Successful      bool      `json:"successful"`
-	Hash            string    `json:"hash"`
-	Ledger          int32     `json:"ledger"`
-	LedgerCloseTime time.Time `json:"created_at"`
-	Account         string    `json:"source_account"`
-	AccountSequence string    `json:"source_account_sequence"`
-	FeeCharged      int32     `json:"fee_charged"`
-	MaxFee          int32     `json:"max_fee"`
-	OperationCount  int32     `json:"operation_count"`
-	EnvelopeXdr     string    `json:"envelope_xdr"`
-	ResultXdr       string    `json:"result_xdr"`
-	ResultMetaXdr   string    `json:"result_meta_xdr"`
-	FeeMetaXdr      string    `json:"fee_meta_xdr"`
-	MemoType        string    `json:"memo_type"`
-	Memo            string    `json:"memo,omitempty"`
-	Signatures      []string  `json:"signatures"`
-	ValidAfter      string    `json:"valid_after,omitempty"`
-	ValidBefore     string    `json:"valid_before,omitempty"`
+	ID                 string              `json:"id"`
+	PT                 string              `json:"paging_token"`
+	Successful         bool                `json:"successful"`
+	Hash               string              `json:"hash"`
+	Ledger             int32               `json:"ledger"`
+	LedgerCloseTime    time.Time           `json:"created_at"`
+	Account            string              `json:"source_account"`
+	AccountSequence    string              `json:"source_account_sequence"`
+	FeeAccount         string              `json:"fee_account"`
+	FeeCharged         int32               `json:"fee_charged"`
+	MaxFee             int32               `json:"max_fee"`
+	OperationCount     int32               `json:"operation_count"`
+	EnvelopeXdr        string              `json:"envelope_xdr"`
+	ResultXdr          string              `json:"result_xdr"`
+	ResultMetaXdr      string              `json:"result_meta_xdr"`
+	FeeMetaXdr         string              `json:"fee_meta_xdr"`
+	MemoType           string              `json:"memo_type"`
+	Memo               string              `json:"memo,omitempty"`
+	Signatures         []string            `json:"signatures"`
+	ValidAfter         string              `json:"valid_after,omitempty"`
+	ValidBefore        string              `json:"valid_before,omitempty"`
+	FeeBumpTransaction *FeeBumpTransaction `json:"fee_bump_transaction,omitempty"`
+	InnerTransaction   *InnerTransaction   `json:"inner_transaction,omitempty"`
+}
+
+// FeeBumpTransaction contains information about a fee bump transaction
+type FeeBumpTransaction struct {
+	Hash       string   `json:"hash"`
+	Signatures []string `json:"signatures"`
+}
+
+// InnerTransaction contains information about the inner transaction contained
+// within a fee bump transaction
+type InnerTransaction struct {
+	Hash       string   `json:"hash"`
+	Signatures []string `json:"signatures"`
+	MaxFee     int32    `json:"max_fee"`
 }
 
 // MarshalJSON implements a custom marshaler for Transaction.
@@ -485,6 +502,9 @@ type TransactionResultCodes struct {
 
 // TransactionSuccess represents the result of a successful transaction
 // submission.
+// Action needed in release: horizonclient-v3.0.0
+// Remove TransactionSuccess because the submit transaction endpoint now responds with
+// a full Transaction resource
 type TransactionSuccess struct {
 	Links struct {
 		Transaction hal.Link `json:"transaction"`