+++ services/horizon/internal/resourceadapter/operations.go
@@ -17,12 +17,13 @@ import (
 func NewOperation(
 	ctx context.Context,
 	operationRow history.Operation,
+	transactionHash string,
 	transactionRow *history.Transaction,
 	ledger history.Ledger,
 ) (result hal.Pageable, err error) {
 
 	base := operations.Base{}
-	PopulateBaseOperation(ctx, &base, operationRow, transactionRow, ledger)
+	PopulateBaseOperation(ctx, &base, operationRow, transactionHash, transactionRow, ledger)
 
 	switch operationRow.Type {
 	case xdr.OperationTypeBumpSequence:
@@ -108,6 +109,7 @@ func PopulateBaseOperation(
 	ctx context.Context,
 	dest *operations.Base,
 	operationRow history.Operation,
+	transactionHash string,
 	transactionRow *history.Transaction,
 	ledger history.Ledger,
 ) {
@@ -122,7 +124,7 @@ func PopulateBaseOperation(
 	dest.SourceAccount = operationRow.SourceAccount
 	populateOperationType(dest, operationRow)
 	dest.LedgerCloseTime = ledger.ClosedAt
-	dest.TransactionHash = operationRow.TransactionHash
+	dest.TransactionHash = transactionHash
 
 	lb := hal.LinkBuilder{Base: httpx.BaseURL(ctx)}
 	self := fmt.Sprintf("/operations/%d", operationRow.ID)
@@ -134,7 +136,7 @@ func PopulateBaseOperation(
 
 	if transactionRow != nil {
 		dest.Transaction = new(horizon.Transaction)
-		PopulateTransaction(ctx, dest.Transaction, *transactionRow)
+		PopulateTransaction(ctx, transactionHash, dest.Transaction, *transactionRow)
 	}
 }