+++ services/horizon/internal/txsub/errors.go
@@ -1,6 +1,7 @@
 package txsub
 
 import (
+	"encoding/hex"
 	"errors"
 	"fmt"
 
@@ -37,12 +38,17 @@ func (fte *FailedTransactionError) Result() (result xdr.TransactionResult, err e
 	return
 }
 
-func (fte *FailedTransactionError) TransactionResultCode() (result string, err error) {
+func (fte *FailedTransactionError) TransactionResultCode(transactionHash string) (result string, err error) {
 	r, err := fte.Result()
 	if err != nil {
 		return
 	}
 
+	innerResult, ok := r.Result.GetInnerResultPair()
+	if ok && transactionHash == hex.EncodeToString(innerResult.TransactionHash[:]) {
+		result, err = codes.String(innerResult.Result.Result.Code)
+		return
+	}
 	result, err = codes.String(r.Result.Code)
 	return
 }
@@ -52,8 +58,7 @@ func (fte *FailedTransactionError) OperationResultCodes() (result []string, err 
 	if err != nil {
 		return
 	}
-
-	oprs, ok := r.Result.GetResults()
+	oprs, ok := r.OperationResults()
 
 	if !ok {
 		return
@@ -69,14 +74,4 @@ func (fte *FailedTransactionError) OperationResultCodes() (result []string, err 
 	}
 
 	return
-}
-
-// MalformedTransactionError represent an error that occurred because
-// a TransactionEnvelope could not be decoded from the provided data.
-type MalformedTransactionError struct {
-	EnvelopeXDR string
-}
-
-func (err *MalformedTransactionError) Error() string {
-	return "tx malformed"
 }