+++ services/horizon/internal/txsub/open_submission_list_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/stellar/go/services/horizon/internal/test"
+	"github.com/stellar/go/support/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 )
@@ -74,10 +75,8 @@ func (suite *SubmissionListTestSuite) TestSubmissionList_Finish() {
 
 	suite.list.Add(suite.ctx, suite.hashes[0], suite.listeners[0])
 	suite.list.Add(suite.ctx, suite.hashes[0], suite.listeners[1])
-	r := Result{
-		Hash: suite.hashes[0],
-	}
-	suite.list.Finish(suite.ctx, r)
+	r := Result{Err: errors.New("test error")}
+	suite.list.Finish(suite.ctx, suite.hashes[0], r)
 
 	// Wries to every listener
 	r1, ok1 := <-suite.listeners[0]
@@ -103,7 +102,7 @@ func (suite *SubmissionListTestSuite) TestSubmissionList_Finish() {
 	assert.False(suite.T(), more)
 
 	// works when no one is waiting for the result
-	err := suite.list.Finish(suite.ctx, r)
+	err := suite.list.Finish(suite.ctx, suite.hashes[0], r)
 	assert.Nil(suite.T(), err)
 }